T2SQL/requirements.txt
ragatouille==0.0.8post2
python-dotenv==1.0.1
openai==1.33.0
inflect==7.3.1
snowflake-connector-python==3.11.0
sqlparse==0.5.1